refactors the three responsibilities of equal/hashcode/signature
Before:
The three responsibilities of equals/hashCode/signature were handled by the same code in SignaturePrinter,
with two major problems:
- no deep equality for package, class and method
- low performance of hashCode

Now:
The three responsibilities are handled by three different visitors:
- SignaturePrinter for signatures
- EqualVisitor for equal
- HashcodeVisitor for equal
- CtElement.getSignature is deprecated
