+++ src/main/java/spoon/reflect/declaration/CtElement.java
@@ -81,7 +81,12 @@ public interface CtElement extends FactoryAccessor, CtVisitable {
 
 	/**
 	 * Gets the signature of the element.
+	 * Has been introduced for CtMethod, see chapter "8.4.2 Method Signature" of the Java specification.
+	 * Overtime, has been badly exploited.
+	 *
+ 	 * Deprecated, will be moved in {@link CtExecutable}, in order to follow the Java specification
 	 */
+	@Deprecated
 	String getSignature();
 
 	/**
